requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
groq==0.13.0
requests==2.32.3
Pillow==10.4.0
python-multipart==0.0.12
gTTS==2.5.1